fix: Atualização tanto do backend como do frontend

+++ models/quote.py
@@ -13,7 +13,7 @@ total_cost: campo calculado (preço * quantidade + frete).
 O engine de custo real usa esse campo para rankings e comparativos.
 """
 
-from sqlalchemy import Column, Integer, Float, String, Date, DateTime, ForeignKey, func
+from sqlalchemy import Column, Integer, Numeric, String, Date, DateTime, ForeignKey, func
 from sqlalchemy.orm import relationship
 
 from core.database import Base
@@ -25,13 +25,13 @@ class Quote(Base):
     id = Column(Integer, primary_key=True, index=True)
     item_id = Column(Integer, ForeignKey("items.id"), nullable=False)
     supplier_id = Column(Integer, ForeignKey("suppliers.id"), nullable=False)
-    unit_price = Column(Float, nullable=False)            # Preço unitário
-    freight = Column(Float, default=0.0)                   # Frete
+    unit_price = Column(Numeric(12, 4), nullable=False)            # Preço unitário
+    freight = Column(Numeric(12, 2), default=0.0)                   # Frete
     delivery_days = Column(Integer, default=7)             # Prazo de entrega (dias)
     payment_term = Column(String(50), default="à vista")   # Condição: à vista, 30 dias, etc
     min_quantity = Column(Integer, default=1)               # Quantidade mínima
     quantity = Column(Integer, default=1)                   # Quantidade cotada
-    total_cost = Column(Float, default=0.0)                # Calculado: (unit_price * qty) + freight
+    total_cost = Column(Numeric(14, 2), default=0.0)                # Calculado: (unit_price * qty) + freight
     valid_until = Column(Date, nullable=True)               # Validade da cotação
     date = Column(Date, nullable=False)                     # Data da cotação
     notes = Column(String(255), nullable=True)
